don't spend time re-building files that have identical hashes
With this patch, crammit will first check if the raw output file exists
before spending time minifying and gzipping the data.  Since the
fingerprint is already in the filename, the output should be identical.

In the vast majority of cases, crammit will only be rebuilding a
fraction of the specified assets.  After the first build, it should now
take significantly less time.