Check for existence of all files before reading.
Before this chance, aborting crammit while it was running could leave
you in an inconsistent state where a concatted file name existed, but
the corresponding minified file did not, which would cause crammit to
error out when you attempted to rebuild.

Now, just rebuild if all the relevant files are not found.